Address PR review comments for stock operations
