+++ lib/data/repository/stock_operations/mock_stock_operations_repository.dart
@@ -242,7 +242,6 @@ class MockStockOperationsRepository extends StockOperationsRepository {
                 )
               : stock,
         )
-        .where((stock) => stock.availableQuantity > 0)
         .toList();
   }
 }
